Fix CountDownTimer handler memory leak
In the original implementation of the class CountDownTimer, the handler
was an anonymous class. In Java, anonymous classes keep an implicit
reference to the outer class.

If we do this:

CountDownTimer countDownTimer = new CountDownTimer(10000, 1000) { ... }.start();
countDownTimer = new CountDownTimer(10000, 1000) { ... }.start();

This causes a memory leak, because when the GC will try to collect the
first instance of CountDownTimer, it will not be able to destroy it, as
the anonymous class keeps a hard reference of the outer class.

This can be fixed creating a static handler class (static classes do
not hold any implicit reference to their outer class). In this case,
we need to create a WeakReference of the outer class, because we need
to acces to some of its properties.

Change-Id: Ibf1d768bde011dbed57bc7522ba8715b8bbcfd01
Signed-off-by: pbr1111 <pbr1111@gmail.com>