[1/2] Add support for pulsing battery icon when charging
This patch adds support for the battery icon to pulse regardless of style.
Commit cba871cbaa5a88609977a45be5d5bdd5df23a270 added a animateSolidBattery method
with the method call guarded under a check for mBatteryStyle being equal to BATTERY_STYLE_SOLID.
We add another check with an OR separator for the STATUS_BAR_PULSE_CHARGING_BATTERY SettingsProvider
to pulse the battery regardless of style when charging.

Also renamed the method to something more descriptive of what we use it for now

Change-Id: I3db110230cdfbabdb446338f6869ca4059cf4f74
Signed-off-by: spezi77 <spezi77@gmx.com>