Multi ScreenRecord tile, blinking stop button, screenrecord service improvements [1/2]
single tap to toggle quality:
- 480*800 resolution and 1.5Mbps
- default resolution (720p) and 4Mbps
- default resolution (720p) and 8Mbps
long press to start recording

Show also a blinking red button at bottom right of the screen
that can be tapped to stop the recording instead of using
the notification.

Also improve the screenrecord service and be sure
to stop it when we are done.

part 2/2 is in framework/av tree, it just sets max recording
duration to 30 minutes instead of 3

Change-Id: I9a0161e6ddafb721421585c7f16bba4b30319425
Signed-off-by: calebcabob <calphonic@gmail.com>