[SQUASH] Advanced screen recording: Ezio edition :P
1. Updated screen recorder using @ezio84's method, in order to work
properly with his power menu:

abc edit: remove PhoneWindowManager power button part,
we already have power+VolUp for partial screenshot so
we're gonna use a tile for this

Removed omni code branding

Change-Id: I252144a4f5c3bdaa84974b996348885ba8dee942

2. Multi ScreenRecord tile, blinking stop button, screenrecord service improvements [1/2]

single tap to toggle quality:
- 480*800 resolution and 1.5Mbps
- default resolution (720p) and 4Mbps
- default resolution (720p) and 8Mbps
long press to start recording

Show also a blinking red button at bottom right of the screen
that can be tapped to stop the recording instead of using
the notification.

Also improve the screenrecord service and be sure
to stop it when we are done.

part 2/2 is in framework/av tree, it just sets max recording
duration to 30 minutes instead of 3

3. Screenrec: let hide the red stop button but keep it clickable

thanks to @KreAch3R for this great idea.

Here we add an option to the screenrecord notification to hide
the red stop hint at bottom right so it doesn't appear in our
recordings, but it's still clickable to stop the service.

4. Screenrecord: show a Share/Delete notification when stopped

also fix MediaScannerConnection memory leak due to aosp bug:
https://issuetracker.google.com/issues/37046656
manually connecting and disconnecting the client

5. Screenrecord red hint: long press on it to move to the opposite side

we can already hide it from notification but we still keep it clickable,
but in some cases we would need to click for example on the keyboard
search/enter button covered by the hint.

Signed-off-by: calebcabob <calphonic@gmail.com>