Disallow extra power menu items when the device is in lock task mode
"Lock task mode" is the mode that the device enters when you pin a app from
recents.

In this mode the user is restricted from using Fling, Pulse and even Smartbar
as well as QS panel. The navbar reverts back to stock and uses the stock
configuration. This is by design. We don't want the user to use any custom
features because this mode is intended to lock a user in an app. They can
obviously exit out of it but that's on them.

With all this said, we want to restrict the power menu items to just the stock
configuration as well. Screenshot, Airplane and Advanced reboot will be hidden
while in this mode.

- We're adapting the original commit to our needs

Original commit -
https://github.com/PixelExperience/frameworks_base/commit/6a4a9e04668b5e62d626341e72a0dc8dcea85a7b

Change-Id: I024a6d6972b9e943d6f3f5fae0b582945dc6519a
Signed-off-by: calebcabob <calphonic@gmail.com>