Configure SystemUI theme [1/2]
In 8.1 several elements change depending on the dominant color of the homescreen
wallpaper. If is a light/white wallpaper these elements will turn white and if is
a dark/black wallpaper then these elements will turn black. If is a neutral color
then it defaults to the light/white color.

This commits makes all of this configurable and so places the option in the user's
hands. They can either leave it up to automatic selection (wallpaper api) or force
a light or dark theme.

-- Commit message was rewritten to better explain what's up :-)

Original commit message:

allow to force light or dark or use the wallpaper based
automatic selection

Signed-off-by: calebcabob <calphonic@gmail.com>