Ensure power menu aligns properly per stock behavior
Setting a scrollview was originally a good idea in the event that
we decided to add more power menu items. The problem is that we're
not planning on adding any more items...ever. The ones we have now
are more than enough.

Another problem with it is that it (scrollview) does not handle
landscape well. So rotating the device sometimes caused items to be
placed incorrectly. Dismissing the power menu and requesting the
layout again would 'fix' this issue but really is not worth that
hassle.

- Move statusbar round corner padding dimens to custom xml
- Remove landscape layout since is no longer necessary
- Expose the padding in case it needs to be adjusted per-device

Change-Id: I7e30e323d631cc987a2fc1c84e191fa86d611191
Signed-off-by: calebcabob <calphonic@gmail.com>