Wallpaper is half black after rotating quickly
When rotate the device quickly in a transparent application activity,
sometime the wallpaper will be half black.

Root cause:
Sometime the transaction that change the size of the wallpaper surface
behind the drawFrame of wallpaper.

Redraw the frame of wallpaper one more time only when wallpaper
size changed can fix it.

Change-Id: I0884d0a6b51691f41e031092a2ef698c7e3a21a5
CRs-Fixed: 2287784
Signed-off-by: Dabug123 <clarkbrunswick55@gmail.com>