Introduce FontTile
In order to accomodate this feature, minor changes were
required to the font service. File permissions on the
font preview files were made more accomodating to allow
other UID's besides system to read them. This should have
been done in the initial font service bring-up, but
wasn't noticed until it became a issue.

In addition, we made the ACCESS_FONT_MANAGER permission
normal to avoid needing users to wipe their devices
to use this feature. A wipe would be necessary from
adding a system level permission to SystemUI manifest.
The service api isn't in the SDK, but if a user app
declared the permission and managed fonts, it would
be interesting at most.

Change-Id: I8b781d11950aaa1e01b7fe47efcaad494bfd3cf5